Fix OCPBUGS-13198
When copying the OCI layout, make sure we only copy files/folders that
are part of the OCI layout specification.

The bug resulted because the copy destination was inside the
source directory, which lead to an infinite circular copy operation when
using the github.com/otiai10/copy package. This fix avoids this
because it copies only the OCI layout files/folders and ignores others.